Update WireGuard FFI in talpid-core
