+++ talpid-core/src/tunnel/wireguard/mod.rs
@@ -15,6 +15,7 @@ use talpid_types::ErrorExt;
 
 pub mod config;
 mod connectivity_check;
+mod logging;
 mod stats;
 pub mod wireguard_go;
 
@@ -133,7 +134,6 @@ impl WireguardMonitor {
         #[cfg(target_os = "windows")]
         route_handle
             .add_default_route_callback(Some(WgGoTunnel::default_route_changed_callback), ());
-
 
         let event_callback = Box::new(on_event.clone());
         let (close_msg_sender, close_msg_receiver) = mpsc::channel();
